refactor: clean up cow amm banner (#5033)
