--- apps/cowswap-frontend/src/common/pure/CoWAMMBanner/cowAmmBannerState.ts
@@ -1,5 +1,0 @@
-import { atom } from 'jotai'
-
-import { StateKey } from './dummyData'
-
-export const cowAmmBannerStateAtom = atom<StateKey>('noLp')
